requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jsonp==1.0.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-renderer-xlsx==0.4.0
et-xmlfile==1.1.0
Markdown==3.3.4
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
